refactor(constant) : ErrorMessage 클래스 이름 변경, 위치 변경
