+++ src/main/java/constant/ErrorMessage.java
@@ -1,6 +1,6 @@
-package domain;
+package constant;
 
-public enum ERROR_MESSAGE {
+public enum ErrorMessage {
     CLOSED_DAY("오늘은 캠퍼스 휴장일입니다."),
     CLOSED_TIME("지금은 캠퍼스 휴장시간입니다."),
     ALREADY_ATTENDED("이미 출석하였습니다. 수정을 원하시면 출석 수정 기능을 이용하세요."),
@@ -12,7 +12,7 @@ public enum ERROR_MESSAGE {
 
     private final String message;
 
-    ERROR_MESSAGE(String message) {
+    ErrorMessage(String message) {
         this.message = message;
     }
