+++ src/main/java/domain/Crew.java
@@ -1,7 +1,7 @@
 package domain;
 
 import static domain.AttendanceStatus.isWeekendOrChristmas;
-import static domain.ERROR_MESSAGE.*;
+import static constant.ErrorMessage.*;
 import static domain.Penalty.COUNSELLING;
 import static domain.Penalty.EXPEL;
 import static domain.Penalty.NONE;
