+++ src/main/java/view/OutputView.java
@@ -1,5 +1,6 @@
 package view;
 
+import domain.Penalty;
 import domain.*;
 
 import java.time.LocalDate;
